refactor: create `shared` directory
