+++ src/main/kotlin/me/daegyeo/netflixchecker/shared/util/EmbedUtil.kt
@@ -1,4 +1,4 @@
-package me.daegyeo.netflixchecker.util
+package me.daegyeo.netflixchecker.shared.util
 
 import discord4j.core.spec.EmbedCreateSpec
 import discord4j.rest.util.Color
